Add FTS5 search, i64 message IDs and persistent name bindings
- Index message text in an FTS5 table kept in sync by triggers and use
  it for search instead of an unindexed LIKE scan over the JSON; user
  queries are normalized to a quoted phrase with prefix match so FTS5
  operators cannot inject syntax. Existing databases are backfilled by
  the (now re-runnable) schema pass on first startup
- Widen message IDs from i32 to i64 throughout the server to match
  SQLite's 64-bit rowids, dropping all try_from conversion boilerplate
- Persist user name to public key bindings in a new user_keys table so
  name/role takeover protection survives restarts; moderation and role
  changes now also resolve offline users via the stored binding, and a
  new `unbind-name` CLI subcommand releases a name after key loss

+++ murmer_server/tests/ephemeral_test.rs
@@ -41,9 +41,7 @@ async fn ephemeral_scan_returns_only_flagged_messages() {
     assert_eq!(rows[0].0, ephemeral_id);
     assert_eq!(rows[0].1, channel);
 
-    let removed = db::delete_message(&db, ephemeral_id as i32)
-        .await
-        .expect("delete");
+    let removed = db::delete_message(&db, ephemeral_id).await.expect("delete");
     assert!(removed);
     assert!(db::get_ephemeral_messages(&db)
         .await
